Add more extensions to create modules. Fix property injections for inline classes.
The injection type of inline classes cannot be retrieved from the method return type, because the signature uses the inner type. So store the generic signature when creating the injection properties (this affects all of them, because there's no way to separate them). Inline classes don't work currently in constructors and methods.